requirements.txt
fastapi==0.112.2
fastapi-cli==0.0.5
entsoe-py==0.6.8
numpy==1.26.4
pandas==2.2.2
python-dotenv==1.0.1
pyarrow==17.0.0
joblib==1.4.2
lightgbm==4.5.0
tqdm==4.66.5
scikit-learn==1.5.2
human-readable==1.3.4